Many changes to allow the source files to compile on arduino devices. Removed all exception handling as not arduino friendly. Removed issue where multiple writes would occur if ACKs were not enabled on pipes.
Note that QoS -1 messages appear to create server errors with missing MIDs, but this is due to -1 QoS messages having no connection to cache the MIDs and isn't important.

The Time.h or TimeLib.h is required to facilitate time functions. Requires Time module install for Arduino.

Tested on Teensy LC and STM32 Bill Pill as MQTT-SN clients